+++ src/main/java/stringcalculator/Operator.java
@@ -1,4 +1,4 @@
-package calculator;
+package stringcalculator;
 
 public interface Operator {
 	long operation(long a, long b);
